pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mpc-mcp-server"
version = "0.3.0"
description = "Metastable Propositional Calculus — thermodynamic framework for resource-bounded inference"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"

dependencies = [
    # ── MCP framework ─────────────────────────────────────────────────────────
    "fastmcp>=2.0.0",

    # ── AI provider SDKs (all five providers supported out-of-the-box) ────────
    "anthropic>=0.45.0",       # Anthropic Claude  (includes models.list())
    "google-generativeai>=0.8.0",  # Google Gemini
    "openai>=1.30.0",          # OpenAI GPT + Kimi (OpenAI-compatible)

    # ── Scientific computing ──────────────────────────────────────────────────
    "numpy>=1.26.0",

    # ── Utilities ─────────────────────────────────────────────────────────────
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# Exact partition function via QuTiP Ising Hamiltonian
qutip = ["qutip>=5.0.0"]

# Spin-glass ground-state solver (NetKet requires JAX)
# Install separately:  pip install "jax[cpu]" netket
netket = ["netket>=3.10.0"]

# Full thermodynamic stack
full = ["mpc[qutip,netket]"]

# Development extras
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
mpc-server = "mpc_server.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["mpc_core*", "mpc_server*"]

[tool.setuptools.package-data]
"*" = ["static/*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
